# Backend/src/exchange_api/green_trade_swagger.json
{
  "swagger": "2.0",
  "info": {
      "description": "API para la plataforma Green Trade. Esta API permite realizar operaciones relacionadas con los usuarios Caficultores, incluyendo el registro y la gestión de sus perfiles.",
      "version": "1.0.0",
      "title": "Green Trade API",
      "contact": {
          "email": "soporte@greentrade.com"
      }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http"],
  "securityDefinitions": {
      "api_key": {
          "type": "apiKey",
          "name": "Authorization",
          "in": "header"
      }
  },
  "paths": {
...
      "caficultor/lot_certifications/post": {
        "post": {
          "tags": ["Caficultor"],
          "summary": "Crea una certificación para un lote específico",
          "description": "Este endpoint registra una nueva certificación para un lote específico, incluyendo el certificado de calidad del producto y el certificado del catador, en la tabla 'lot_certifications'.",
          "operationId": "createLotCertification",
          "consumes": ["multipart/form-data"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "formData",
              "name": "id_lot",
              "type": "integer",
              "required": true,
              "description": "ID del lote al que pertenece la certificación."
            },
            {
              "in": "formData",
              "name": "product_sc_certificate",
              "type": "file",
              "required": true,
              "description": "Archivo del certificado de calidad del producto."
            },
            {
              "in": "formData",
              "name": "product_taster_certificate",
              "type": "file",
              "required": true,
              "description": "Archivo del certificado del catador."
            },
            {
              "in": "formData",
              "name": "contact_qgrader",
              "type": "boolean",
              "required": true,
              "description": "Indica si se incluye información de contacto de un Q-Grader."
            }
          ],
          "responses": {
            "201": {
              "description": "Certificación del lote creada con éxito."
            },
            "400": {
              "description": "Datos faltantes o inválidos en la solicitud."
            },
            "404": {
              "description": "Lote no encontrado."
            },
            "500": {
              "description": "Error interno del servidor."
            }
          }
        }
      },
...
      "caficultor/user_role": {
        "post": {
          "tags": ["Caficultor"],
          "summary": "Asigna un rol a un usuario",
          "description": "Este endpoint permite asignar un nuevo rol a un usuario. Guarda la asignación en la tabla 'user_role'.",
          "operationId": "addNewUserRole",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "id_user",
              "in": "formData",
              "required": true,
              "type": "integer",
              "description": "El ID del usuario al que se le asignará el rol."
            },
            {
              "name": "id_role",
              "in": "formData",
              "required": true,
              "type": "integer",
              "description": "El ID del rol que se asignará al usuario."
            }
          ],
          "responses": {
            "201": {
              "description": "Rol de usuario agregado exitosamente."
            },
            "400": {
              "description": "Error de validación."
            },
            "404": {
              "description": "Usuario no encontrado."
            },
            "500": {
              "description": "Error interno del servidor."
            }
          }
        }
      },
...
      "/administrador/cursos/get/{id_curso}": {
        "get": {
            "tags": ["Administrador"],
            "summary": "Obtener detalles de un curso",
            "description": "Endpoint para obtener los detalles de un curso por su ID.",
            "operationId": "getCourseById",
            "produces": ["application/json"],
            "parameters": [
                {
                    "name": "id_curso",
                    "in": "path",
                    "required": true,
                    "type": "integer",
                    "description": "ID único del curso a obtener."
                }
            ],
            "responses": {
                "200": {
                    "description": "Detalles del curso obtenidos con éxito."
                },
                "400": {
                    "description": "ID de curso inválido."
                },
                "404": {
                    "description": "Curso no encontrado."
                },
                "500": {
                    "description": "Error interno del servidor."
                }
            }
        }
      },
...
  },
...
}